completed authorization, checks for duplicate emails before account c…

+++ controllers/quiz.js
@@ -3,14 +3,13 @@ const quizzes = express.Router();
 const Quiz = require('../models/quiz');
 const seedData = require('../models/seedModel');
 
-//index (quizzes list)
+// index (quizzes list)
 quizzes.get('/', (req, res) => {
 	Quiz.find({}, (err, allQuizzes) => {
 		let quizArray = [];
 		for (i = 0; i < allQuizzes.length; i++) {
 			let quizObject = {
 				name: allQuizzes[i].name,
-				// caption: allQuizzes[i].caption,
 				image: allQuizzes[i].image,
 				createdBy: allQuizzes[i].createdBy,
 				createdAt: allQuizzes[i].created_at,
@@ -26,34 +25,28 @@ quizzes.get('/', (req, res) => {
 	});
 });
 
-// create
+// create quiz
 quizzes.post('/', (req, res) => {
-	const cookies = parseCookies(req.headers.cookie);
-	const userObjectID = cookies.userid;
-	const updatedQuiz = req.body;
-	Object.assign(updatedQuiz, { createdBy: userObjectID });
-	// console.log('CREATED QUIZ BELOW')
-	// console.log(updatedQuiz);
+	// Check if current session exists before posting
+	if (req.session.currentUser) {
+		const cookies = parseCookies(req.headers.cookie);
+		const userObjectID = cookies.userid;
+		const updatedQuiz = req.body;
+		Object.assign(updatedQuiz, { createdBy: userObjectID });
 
-	Quiz.create(updatedQuiz, (err, createdQuiz) => {
-		if (err) {
-			res.status(400).json({ error: err.message });
-		} else {
-			res.status(200).send(createdQuiz);
-		}
-	});
+		Quiz.create(updatedQuiz, (err, createdQuiz) => {
+			if (err) {
+				res.status(400).send(err);
+			} else {
+				res.status(200).send(createdQuiz);
+			}
+		});
+	} else {
+		res.status(400).send('Invalid session, try logging out and in again.');
+	}
 });
 
-quizzes.get('/', (req, res) => {
-	//auth logic here
-	// do an axios call to this route
-	// console.log(req.headers.cookie);
-	// console.log(req.session);
-	// console.log('test');
-	res.send('You have found The Knowledges backend!');
-});
-
-//MAKE THE PUT ROUTE FOR THE ITERATOR
+// update quiz count
 quizzes.put('/:id', (req, res) => {
 	Quiz.findByIdAndUpdate(
 		req.params.id,
@@ -68,35 +61,33 @@ quizzes.put('/:id', (req, res) => {
 	);
 });
 
-// delete
+// delete quiz
 quizzes.delete('/:id', (req, res) => {
-	Quiz.findByIdAndRemove(req.params.id, (err, deletedQuiz) => {
-		if (err) {
-			res.status(400).json({ error: err.message });
-		}
-		res.status(200).json(deletedQuiz);
-	});
+	// Check if current session exists before posting
+	if (req.session.currentUser) {
+		Quiz.findByIdAndRemove(req.params.id, (err, deletedQuiz) => {
+			if (err) {
+				res.status(400).json({ error: err.message });
+			}
+			res.status(200).json(deletedQuiz);
+		});
+	} else {
+		res.status(400).send('Invalid session, try logging out and in again.');
+	}
 });
 
-//MAKE a route that pulls all quiz information from a specific quiz
-//create a get route in the quiz controller that pulls all data from the
-//json object of quiz information
-//res.send to call all the info
-
-//seed
+// seed quizzes
 quizzes.get('/seed', (req, res) => {
-	console.log('Attempting to seed');
 	Quiz.create(seedData, (err, createdQuizzes) => {
 		if (err) {
 			res.status(400).json({ error: err.message });
 		} else {
-			console.log('Successfully seeded data');
 			res.send('Data successfully seeded');
 		}
 	});
 });
 
-//Quiz info pull
+// get individual quiz
 quizzes.get('/:id', (req, res) => {
 	Quiz.findById(req.params.id, (err, quizInfo) => {
 		if (err) {
@@ -106,8 +97,6 @@ quizzes.get('/:id', (req, res) => {
 		}
 	});
 });
-
-module.exports = quizzes;
 
 // Adapted from https://stackoverflow.com/questions/3393854/
 const parseCookies = cookies => {
@@ -120,3 +109,5 @@ const parseCookies = cookies => {
 
 	return cookiesObj;
 };
+
+module.exports = quizzes;
